requirements.txt
numpy
pillow
open3d
gymnasium
mani_skill
pytest
torch
transformers
open_clip_torch
